challenge 10, playlist.py, playlist functions and library.py
