+++ Mianta/10-challenge/playlist_functions.py
@@ -1,7 +1,9 @@
 # NOTE: these first two functions do not have to be edited
 
 # prints out what is in your playlist
-# takes one argument: 'playlist' (a list)
+# takes one argument: 'playlist' (a list)NOTE: this prints everything on the playlist and is a parameter
+from turtle import title
+
 def display_playlist(playlist):
 	if len(playlist) == 0:
 		print('Playlist is empty!')
@@ -17,32 +19,19 @@ def add_song(playlist, song):
 	song['plays'] = 0
 	playlist.append(song)
 
-
-
+# This function should have one parameter called 'playlist'
 '''
-TODO (Question 6): define a function called get_playlist_length()
-This function should have one parameter called 'playlist'
 The function should return an integer value indicating how many songs there are
 '''
+def get_playlist_length(playlist):#TODO (Question 6):define a function called get_playlist_length()
+	return len(playlist)
+# #TODO (BONUS): define a function called play_track()
+# It should have two parameters-'playlist' (a list)-'track' (an integer) - this should be optional, and by default play track 1
 
+#This function should 'play' the song corresponding to the input track 
 
-
-
-'''
-TODO (BONUS): define a function called play_track()
-It should have two parameters
--'playlist' (a list)
--'track' (an integer) - this should be optional, and by default play track 1
-
-This function should 'play' the song corresponding to the input track #
-
-For example play_track(my_playlist,3) should print out:
-
-'Now playing track 3, Controversy by Prince' 
-Assuming that the third track in your playlist 'Controversy' by 'Prince'
-
-This function should ALSO increase the 'plays' value for that song's dictionary by 1
-So, if 'Controversy' has 0 plays so far, it should now be increased to 1
-'''
-
-
+def play_track(playlist, track = 1):
+	for i in range(len(playlist)):
+		if track == i+1:
+			print(f'Now playing track {track}, {playlist[i]["title"]} by {playlist[i]["artist"]}')
+			playlist[i]['plays'] += 1
